chore: code cleanup + documentation
