+++ src/xcode/ENA/ENA/Source/Extensions/NSMutableAttributedString+Generation.swift
@@ -1,0 +1,40 @@
+//
+// Corona-Warn-App
+//
+// SAP SE and all other contributors /
+// copyright owners license this file to you under the Apache
+// License, Version 2.0 (the "License"); you may not use this
+// file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
+//
+
+import Foundation
+
+
+extension NSMutableAttributedString {
+
+	/// Generates an attributed string from a normal text + attributed text injected into this text.
+	static func generateAttributedString(normalText: String, attributedText: [NSAttributedString]) -> NSMutableAttributedString {
+
+		let components = normalText.components(separatedBy: "%@")
+		let adjusted: NSMutableAttributedString = NSMutableAttributedString(string: "")
+
+		for (index, element) in components.enumerated() {
+			adjusted.append(NSAttributedString(string: element))
+			if index < attributedText.count {
+				adjusted.append(attributedText[index])
+			}
+		}
+
+		return adjusted
+	}
+}
